requirements.txt
neo4j>=5.15.0
streamlit>=1.29.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
ollama>=0.1.7
pyvis>=0.3.2
scikit-learn>=1.3.0
matbench>=0.7.0
mp-api>=0.35.0
pathlib>=1.0.1
